Move all contract results in results directory and split by type
