+++ packages/std/src/results/init.rs
@@ -1,0 +1,71 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+use std::fmt;
+
+use crate::errors::StdResult;
+use crate::types::Empty;
+
+use super::attribute::Attribute;
+use super::cosmos_msg::CosmosMsg;
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct InitResponse<T = Empty>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    pub messages: Vec<CosmosMsg<T>>,
+    /// The attributes that will be emitted as part of a "wasm" event
+    pub attributes: Vec<Attribute>,
+}
+
+pub type InitResult<U = Empty> = StdResult<InitResponse<U>>;
+
+impl<T> Default for InitResponse<T>
+where
+    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    fn default() -> Self {
+        InitResponse {
+            messages: vec![],
+            attributes: vec![],
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::super::BankMsg;
+    use super::*;
+    use crate::addresses::HumanAddr;
+    use crate::errors::StdError;
+    use crate::{coins, from_slice, to_vec};
+
+    #[test]
+    fn can_deser_error_result() {
+        let fail = InitResult::Err(StdError::Unauthorized { backtrace: None });
+        let bin = to_vec(&fail).expect("encode contract result");
+        println!("error: {}", std::str::from_utf8(&bin).unwrap());
+        let back: InitResult = from_slice(&bin).expect("decode contract result");
+        assert_eq!(fail, back);
+    }
+
+    #[test]
+    fn can_deser_ok_result() {
+        let send = InitResult::Ok(InitResponse {
+            messages: vec![BankMsg::Send {
+                from_address: HumanAddr::from("me"),
+                to_address: HumanAddr::from("you"),
+                amount: coins(1015, "earth"),
+            }
+            .into()],
+            attributes: vec![Attribute {
+                key: "action".to_string(),
+                value: "release".to_string(),
+            }],
+        });
+        let bin = to_vec(&send).expect("encode contract result");
+        println!("ok: {}", std::str::from_utf8(&bin).unwrap());
+        let back: InitResult = from_slice(&bin).expect("decode contract result");
+        assert_eq!(send, back);
+    }
+}
